+++ packages/svelte/src/compiler/phases/3-transform/server/visitors/EachBlock.js
@@ -1,9 +1,8 @@
 /** @import { BlockStatement, Expression, Statement } from 'estree' */
 /** @import { AST } from '#compiler' */
 /** @import { ComponentContext } from '../types.js' */
-import { BLOCK_OPEN_ELSE } from '../../../../../internal/server/hydration.js';
 import * as b from '#compiler/builders';
-import { block_close, block_open } from './shared/utils.js';
+import { block_close, block_open, block_open_else, create_async_block } from './shared/utils.js';
 
 /**
  * @param {AST.EachBlock} node
@@ -18,7 +17,9 @@ export function EachBlock(node, context) {
 		each_node_meta.contains_group_binding || !node.index ? each_node_meta.index : b.id(node.index);
 
 	const array_id = state.scope.root.unique('each_array');
-	state.init.push(b.const(array_id, b.call('$.ensure_array_like', collection)));
+
+	/** @type {Statement} */
+	let block = b.block([b.const(array_id, b.call('$.ensure_array_like', collection))]);
 
 	/** @type {Statement[]} */
 	const each = [];
@@ -44,23 +45,27 @@ export function EachBlock(node, context) {
 	);
 
 	if (node.fallback) {
-		const open = b.stmt(b.call(b.member(b.id('$$payload.out'), b.id('push')), block_open));
+		const open = b.stmt(b.call(b.id('$$renderer.push'), block_open));
 
 		const fallback = /** @type {BlockStatement} */ (context.visit(node.fallback));
 
-		fallback.body.unshift(
-			b.stmt(b.call(b.member(b.id('$$payload.out'), b.id('push')), b.literal(BLOCK_OPEN_ELSE)))
-		);
+		fallback.body.unshift(b.stmt(b.call(b.id('$$renderer.push'), block_open_else)));
 
-		state.template.push(
+		block.body.push(
 			b.if(
 				b.binary('!==', b.member(array_id, 'length'), b.literal(0)),
 				b.block([open, for_loop]),
 				fallback
-			),
-			block_close
+			)
 		);
 	} else {
-		state.template.push(block_open, for_loop, block_close);
+		state.template.push(block_open);
+		block.body.push(for_loop);
+	}
+
+	if (node.metadata.expression.has_await) {
+		state.template.push(create_async_block(block), block_close);
+	} else {
+		state.template.push(...block.body, block_close);
 	}
 }