+++ packages/svelte/src/compiler/phases/3-transform/server/visitors/SvelteBoundary.js
@@ -1,17 +1,14 @@
 /** @import { BlockStatement } from 'estree' */
 /** @import { AST } from '#compiler' */
 /** @import { ComponentContext } from '../types' */
-import { BLOCK_CLOSE, BLOCK_OPEN } from '../../../../../internal/server/hydration.js';
 import * as b from '#compiler/builders';
-import { build_attribute_value } from './shared/utils.js';
+import { block_close, block_open, block_open_else, build_attribute_value } from './shared/utils.js';
 
 /**
  * @param {AST.SvelteBoundary} node
  * @param {ComponentContext} context
  */
 export function SvelteBoundary(node, context) {
-	context.state.template.push(b.literal(BLOCK_OPEN));
-
 	// if this has a `pending` snippet, render it
 	const pending_attribute = /** @type {AST.Attribute} */ (
 		node.attributes.find((node) => node.type === 'Attribute' && node.name === 'pending')
@@ -23,16 +20,23 @@ export function SvelteBoundary(node, context) {
 		)
 	);
 
-	if (pending_attribute) {
-		const value = build_attribute_value(pending_attribute.value, context, false, true);
-		context.state.template.push(b.call(value, b.id('$$payload')));
-	} else if (pending_snippet) {
-		context.state.template.push(
-			/** @type {BlockStatement} */ (context.visit(pending_snippet.body))
-		);
-	} else {
-		context.state.template.push(/** @type {BlockStatement} */ (context.visit(node.fragment)));
-	}
+	if (pending_attribute || pending_snippet) {
+		const pending = pending_attribute
+			? b.call(
+					build_attribute_value(
+						pending_attribute.value,
+						context,
+						(expression) => expression,
+						false,
+						true
+					),
+					b.id('$$renderer')
+				)
+			: /** @type {BlockStatement} */ (context.visit(pending_snippet.body));
 
-	context.state.template.push(b.literal(BLOCK_CLOSE));
+		context.state.template.push(block_open_else, pending, block_close);
+	} else {
+		const block = /** @type {BlockStatement} */ (context.visit(node.fragment));
+		context.state.template.push(block_open, block, block_close);
+	}
 }