+++ packages/svelte/src/compiler/phases/types.d.ts
@@ -1,10 +1,10 @@
 import type { AST, Binding, StateField } from '#compiler';
 import type {
+	AwaitExpression,
 	CallExpression,
 	ClassBody,
 	Identifier,
 	LabeledStatement,
-	Node,
 	Program
 } from 'estree';
 import type { Scope, ScopeRoot } from './scope.js';
@@ -47,6 +47,8 @@ export interface Analysis {
 
 	/** A set of deriveds that contain `await` expressions */
 	async_deriveds: Set<CallExpression>;
+	/** Awaits needing context preservation */
+	pickled_awaits: Set<AwaitExpression>;
 }
 
 export interface ComponentAnalysis extends Analysis {