+++ packages/svelte/src/compiler/utils/ast.js
@@ -609,3 +609,27 @@ export function build_assignment_value(operator, left, right) {
 			? b.logical(/** @type {ESTree.LogicalOperator} */ (operator.slice(0, -1)), left, right)
 			: b.binary(/** @type {ESTree.BinaryOperator} */ (operator.slice(0, -1)), left, right);
 }
+
+/**
+ * @param {ESTree.Expression} expression
+ */
+export function has_await(expression) {
+	let has_await = false;
+
+	walk(expression, null, {
+		AwaitExpression(_node, context) {
+			has_await = true;
+			context.stop();
+		}
+	});
+
+	return has_await;
+}
+
+/**
+ * Turns `await ...` to `(await $.save(...))()`
+ * @param {ESTree.Expression} expression
+ */
+export function save(expression) {
+	return b.call(b.await(b.call('$.save', expression)));
+}