+++ packages/svelte/tests/runtime-runes/samples/async-no-pending-await-in-script-2/main.svelte
@@ -1,0 +1,9 @@
+<script>
+	import Child from './Child.svelte';
+</script>
+
+<Child />
+
+{#if true}
+	<p>hello</p>
+{/if}